fix: include LATE status in salary calculation and update security rules

+++ src/main/java/com/mytraxo/payroll/entity/PayrollRecord.java
@@ -23,7 +23,7 @@ public class PayrollRecord {
     private int year;
 
     private Integer totalDaysInMonth;
-    private Long presentDays;
+    private Double presentDays;
 
     private Double basicPay;
     private Double hra;
